Bring back telegraf, influxdb, and grafana. (#790)
* Remove clientid from help message and keeps track of requested tasks.

* Bring back telegraf, influxdb, and grafana

+++ routerlicious/src/agent/metricClient.ts
@@ -30,9 +30,9 @@ class TelegrafClient implements IMetricClient {
 
     private createTelegrafRow(traces: core.ITrace[]): Object {
         const row = new Object();
-        const Int = telegraf.Int;
+        const Float = telegraf.Float;
         for (const trace of traces) {
-            row[trace.service + "-" + trace.action] = new Int(trace.timestamp);
+            row[trace.service + "-" + trace.action] = new Float(trace.timestamp);
         }
         return row;
     }
